backend/requirements.txt
Django>=5.0.3
djangorestframework>=3.15.1
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
celery>=5.3.6
redis>=5.0.3
Pillow>=10.3.0
django-mptt>=0.16.0
django-import-export>=3.3.8
openpyxl>=3.1.2
pandas>=2.2.1
cloudinary>=1.39.0
requests>=2.31.0
firebase-admin>=6.5.0
django-celery-results>=2.5.1
django-celery-beat>=2.6.0
python-dotenv>=1.0.1
